requirements.txt
fastapi>=0.95.0
uvicorn>=0.21.1
pydantic>=1.10.7
email-validator>=2.0.0
sqlalchemy>=2.0.9
passlib>=1.7.4
python-jose>=3.3.0
bcrypt>=4.0.1
python-multipart>=0.0.6
pillow>=9.5.0
psutil>=5.9.5
python-dotenv>=1.0.0
httpx>=0.24.0
websockets>=11.0.2
psycopg2-binary>=2.9.5 